Move definitions for Conduction
