--- Conduction/Conduction.csproj
@@ -1,14 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
-    <AssemblyName>YggdrAshill.Nuadha.Conduction</AssemblyName>
-    <RootNamespace>YggdrAshill.Nuadha.Conduction</RootNamespace>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Signalization\Signalization.csproj" />
-    <ProjectReference Include="..\Unitization\Unitization.csproj" />
-  </ItemGroup>
-
-</Project>
