+++ Signalization/IEmission.cs
@@ -1,4 +1,4 @@
-namespace YggdrAshill.Nuadha.Conduction
+namespace YggdrAshill.Nuadha.Signalization
 {
     /// <summary>
     /// Token to <see cref="Emit"/>.
