+++ YggdrAshill.Nuadha.Specification/Signalization/GenerateSpecification.cs
