+++ YggdrAshill.Nuadha.Specification/Signalization/PropagateSpecification.cs
@@ -1,6 +1,5 @@
 using NUnit.Framework;
 using YggdrAshill.Nuadha.Signalization;
-using YggdrAshill.Nuadha.Conduction;
 using System;
 
 namespace YggdrAshill.Nuadha.Specification
