+++ YggdrAshill.Nuadha/Signalization/Emission.cs
@@ -1,4 +1,4 @@
-using YggdrAshill.Nuadha.Conduction;
+using YggdrAshill.Nuadha.Signalization;
 using System;
 
 namespace YggdrAshill.Nuadha
