+++ YggdrAshill.Nuadha/Signalization/Propagate.cs
@@ -1,5 +1,4 @@
 using YggdrAshill.Nuadha.Signalization;
-using YggdrAshill.Nuadha.Conduction;
 using System;
 using System.Collections.Generic;
