+++ YggdrAshill.Nuadha/Signals/Space3DPositionTo.cs
@@ -1,5 +1,5 @@
+using YggdrAshill.Nuadha.Signalization;
 using YggdrAshill.Nuadha.Transformation;
-using YggdrAshill.Nuadha.Conduction;
 using YggdrAshill.Nuadha.Signals;
 using System;
