+++ YggdrAshill.Nuadha/Transformation/SignalTo.cs
@@ -1,6 +1,5 @@
 using YggdrAshill.Nuadha.Signalization;
 using YggdrAshill.Nuadha.Transformation;
-using YggdrAshill.Nuadha.Conduction;
 using System;
 
 namespace YggdrAshill.Nuadha
