+++ YggdrAshill.Nuadha/Units/Stick/IStickConfiguration.cs
@@ -1,4 +1,4 @@
-using YggdrAshill.Nuadha.Conduction;
+using YggdrAshill.Nuadha.Signalization;
 using YggdrAshill.Nuadha.Signals;
 
 namespace YggdrAshill.Nuadha
