+++ YggdrAshill.Nuadha/Units/Transmit.cs
@@ -1,6 +1,5 @@
 using YggdrAshill.Nuadha.Signalization;
 using YggdrAshill.Nuadha.Unitization;
-using YggdrAshill.Nuadha.Conduction;
 using YggdrAshill.Nuadha.Units;
 using System;
