+++ YggdrAshill.Nuadha/YggdrAshill.Nuadha.csproj
@@ -5,7 +5,6 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\Conduction\Conduction.csproj" />
     <ProjectReference Include="..\Signalization\Signalization.csproj" />
     <ProjectReference Include="..\Signals\Signals.csproj" />
     <ProjectReference Include="..\Transformation\Transformation.csproj" />
